one to many set between student and user

+++ src/main/java/com/example/FirstApp/service/CourseService.java
@@ -6,6 +6,9 @@ import com.example.FirstApp.entity.Course;
 import com.example.FirstApp.repository.CourseRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -18,26 +21,34 @@ public class CourseService {
     @Autowired
     private CourseRepository courseRepository;
 
-    public CourseDTO addCourse(Course course){
+    public ResponseEntity<?> addCourse(Course course){
+        if(course.getCourseName() ==  null){
+            return new ResponseEntity<>("course name required", HttpStatus.BAD_REQUEST);
+        }
+
+        if(course.getCourseDescription() ==  null){
+            return new ResponseEntity<>("course Description required", HttpStatus.BAD_REQUEST);
+        }
+
         CourseDTO dto = new CourseDTO();
         dto.setCourseName(course.getCourseName());
-        dto.setAddedDate(course.getAddedDate());
+        dto.setCourseDes(course.getCourseDescription());
 
         courseRepository.save(course);
-
-        return dto;
+        return new ResponseEntity<>(dto,HttpStatus.CREATED);
     }
 
-    public CourseDTO getCourseById(Long id){
-        CourseDTO courseDTO = new CourseDTO();
+    public ResponseEntity<?> getCourseById(Long id){
+       return courseRepository.findById(id)
+                        .map(course -> {
+                                    CourseDTO courseDTO = new CourseDTO();
+                                    courseDTO.setCourseDes(course.getCourseDescription());
+                                    courseDTO.setCourseName(course.getCourseName());
 
-        Course course = courseRepository.findById(id).orElse(new Course());
-        if(course.getCourseId() == null)
-            return new CourseDTO();
-
-        courseDTO.setAddedDate(course.getAddedDate());
-        courseDTO.setCourseName(course.getCourseName());
-        return courseDTO;
+                                    return ResponseEntity.ok(courseDTO);
+                                }
+                        )
+                                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
     public List<CourseDTO> getAllCourses(){
@@ -47,27 +58,39 @@ public class CourseService {
         for(Course x : courses) {
             CourseDTO courseDTO = new CourseDTO();
             courseDTO.setCourseName(x.getCourseName());
-            courseDTO.setAddedDate(x.getAddedDate());
+            courseDTO.setCourseDes(x.getCourseDescription());
 
             result.add(courseDTO);
         }
         return result;
     }
 
-    public List<CourseDTO> deleteCourseById(Long id){
+    public ResponseEntity<?>  deleteCourseById(Long id){
+        ResponseEntity<?> response = getCourseById(id);
+
+        if(response.getStatusCode().equals(HttpStatus.NOT_FOUND))
+            return new ResponseEntity<>("No such record",HttpStatus.NOT_FOUND);
+
         courseRepository.deleteById(id);
-        return getAllCourses();
+        return new ResponseEntity<>("Course deleted",HttpStatus.ACCEPTED);
     }
 
-    public void updateCourseById(Long id, Course course){
-        Course c = courseRepository.findById(id).orElse(new Course());
+    public  ResponseEntity<?> updateCourseById(Long id, Course course){
+        Course c = courseRepository.findById(id).orElseGet(null);
 
-        if(c.getCourseId() == null)
-            return;
+        if(c == null)
+            return new ResponseEntity<>("No such record",HttpStatus.NOT_FOUND);
 
-        c.setCourseName(course.getCourseName());
-        c.setAddedDate(course.getAddedDate());
+        if(course.getCourseName() != null)
+            c.setCourseName(course.getCourseName());
+
+        if(course.getCourseDescription() != null)
+            c.setCourseDescription(course.getCourseDescription());
+
         courseRepository.save(c);
+        CourseDTO courseDTO = new CourseDTO(c.getCourseName(), c.getCourseDescription());
+
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(courseDTO);
     }
