+++ src/main/java/com/example/FirstApp/service/StudentService.java
@@ -2,15 +2,19 @@ package com.example.FirstApp.service;
 
 
 import com.example.FirstApp.DTO.CourseDTO;
-import com.example.FirstApp.entity.Course;
-import com.example.FirstApp.entity.Student;
+import com.example.FirstApp.entity.*;
 import com.example.FirstApp.repository.CourseRepository;
 import com.example.FirstApp.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -18,38 +22,80 @@ public class StudentService {
     @Autowired
     private StudentRepository studentRepository;
 
-    public Student addStudent(Student student){
-        return studentRepository.save(student);
+    @Autowired
+    private UserService userService;
+
+    public ResponseEntity<?> addStudent(Student student, Long userId){
+        User user = userService.getUser(userId);
+
+        System.out.println(user);
+
+        if(user == null || student == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("No such User exist");
+
+        if(student.getLastName() == null || student.getFirstName() == null || student.getEmailId() == null){
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Data Not sufficient");
+        }
+
+        Guardian guardian = student.getGuardian();
+        if(guardian == null)
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("guardian details missing");
+
+
+        if(guardian.getEmail() == null || guardian.getMobile() == null || guardian.getName() == null){
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Guardian Details missing");
+        }
+
+        student.setAddedBY(user);
+        studentRepository.save(student);
+
+        return new ResponseEntity<>(student,HttpStatus.CREATED);
     }
 
-    public Student getStudentById(Long id){
-        return studentRepository.findById(id).orElse(new Student());
+    public ResponseEntity<?> getStudentById(Long id){
+        return studentRepository.findById(id)
+                .map(ResponseEntity :: ok)
+                .orElseGet(() -> (ResponseEntity<Student>) notFound());
     }
 
     public List<Student> getAllStudents(){
-        List<Student> courses = studentRepository.findAll();
-        return courses;
+        List<Student> students = studentRepository.findAll();
+        return students;
     }
 
-    public List<Student> deleteStudentById(Long id){
+    public ResponseEntity<?> deleteStudentById(Long id){
+        Student student = studentRepository.findById(id).orElse(null);
+        if(student == null)
+            return notFound();
+
         studentRepository.deleteById(id);
-        return getAllStudents();
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Deleted");
     }
 
-    public void updateStudentById(Long id, Student student){
-        Student c = studentRepository.findById(id).orElse(new Student());
+    public ResponseEntity<?> notFound(){
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No such data");
+    }
 
-        if(c.getStudentId() == null)
-            return;
+    public ResponseEntity<?> updateStudentById(Long id, Student student){
+        Student c = studentRepository.findById(id).orElse(null);
+        if(c == null)
+            return notFound();
 
-        c.setFirstName(student.getFirstName());
-        c.setLastName(student.getLastName());
-        c.setAddedDate(student.getAddedDate());
-        c.setEmailId(student.getEmailId());
-        c.setGuardian(student.getGuardian());
-        c.setCourseList(student.getCourseList());
+        if(student.getFirstName() != null)
+            c.setFirstName(student.getFirstName());
+
+        if(c.getLastName() != null)
+            c.setLastName(student.getLastName());
+
+        if(c.getEmailId() != null)
+            c.setEmailId(student.getEmailId());
+
+        if(student.getGuardian() != null)
+            c.setGuardian(student.getGuardian());
+
 
         studentRepository.save(c);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Updated");
     }
 
 }
